Tags class created for the sole purpose of changing tags and such; currently only populated with the location to save the QR code into the internal memory in the pictures folder.
RetrievedFromAPI class created extending AsyncTask, allowing API to be accessed on another thread. This contains the API code, including the key which needs to get moved to a different location but for right now it's fine, to push the request to the API, as well as processing the QR code received into a Bitmap. This bitmap file is a private field with a getter function.

Also includes DashboardFragment additions to onCreate() so that the new thread can run when the "Create QR Code" button is pressed.